+++ Core/Interfaces/IMasterServicesService.cs
@@ -5,7 +5,7 @@ namespace BusinessLogic.Interfaces;
 public interface IMasterServicesService
 {
     public void AddMasterServiceFromList(int idMaster, int idProvidedService);
-    void AddService(string name, decimal price);
+    void AddService(string name, float price);
     void DeleteService(int idMaster, int idService);
     public List<MasterService> GetMasterServices(int masterId);
     public MasterServicesAndMasterDto GetSevicesToDisplayForMaster(int? masterId);
